Add empty params move GetUidMin
