+++ src/modules/complianceengine/src/lib/Users.h
@@ -1,0 +1,6 @@
+#include <ContextInterface.h>
+
+namespace ComplianceEngine
+{
+int GetUidMin(ContextInterface& context);
+} // namespace ComplianceEngine
