+++ src/modules/complianceengine/src/lib/procedures/EnsureAccountsWithoutShellAreLocked.h
@@ -5,9 +5,20 @@
 #define COMPLIANCEENGINE_PROCEDURES_ENSURE_ACCOUNTS_WITHOUT_SHELL_ARE_LOCKED_H
 
 #include <Evaluator.h>
+#include <Separated.h>
 
 namespace ComplianceEngine
 {
-Result<Status> AuditEnsureAccountsWithoutShellAreLocked(IndicatorsTree& indicators, ContextInterface& context);
+
+struct AuditEnsureAccountsWithoutShellAreLockedParams
+{
+
+    /// List of users to be excluded from check
+    Separated<std::string, ':'> excludeUsers;
+    /// Parse /etc/login.defs and skip users with uid below UID_MIN
+    bool skip_below_uid_min = false;
+};
+Result<Status> AuditEnsureAccountsWithoutShellAreLocked(const AuditEnsureAccountsWithoutShellAreLockedParams& params, IndicatorsTree& indicators,
+    ContextInterface& context);
 } // namespace ComplianceEngine
 #endif // COMPLIANCEENGINE_PROCEDURES_ENSURE_ACCOUNTS_WITHOUT_SHELL_ARE_LOCKED_H
