+++ src/modules/complianceengine/src/lib/procedures/EnsureAccountsWithoutShellAreLocked.schema.json
@@ -12,9 +12,21 @@
           "properties": {
             "EnsureAccountsWithoutShellAreLocked": {
               "type": "object",
-              "required": [],
+              "required": [
+                "excludeUsers",
+                "skip_below_uid_min"
+              ],
               "additionalProperties": false,
-              "properties": {}
+              "properties": {
+                "excludeUsers": {
+                  "type": "string",
+                  "description": "List of users to be excluded from check"
+                },
+                "skip_below_uid_min": {
+                  "type": "string",
+                  "description": "Parse /etc/login.defs and skip users with uid below UID_MIN"
+                }
+              }
             }
           }
         }
